organize method channel calls
