+++ lib/src/livekit.dart
@@ -1,6 +1,3 @@
-import 'package:flutter/services.dart';
-import 'package:meta/meta.dart';
-
 import 'core/room.dart';
 import 'options.dart';
 
@@ -8,9 +5,6 @@ import 'options.dart';
 /// {@category Room}
 class LiveKitClient {
   static const version = '0.5.6';
-
-  @internal
-  static const channel = MethodChannel('livekit_client');
 
   /// Convenience method for connecting to a LiveKit server.
   /// Returns a [Room] upon a successful connect or throws when it fails.
